Update Langchain version
