+++ app/qna/db.py
@@ -1,0 +1,49 @@
+from redis import Redis
+from typing import List
+
+from langchain.schema import Document
+from langchain.vectorstores.redis import Redis as RedisVDB
+
+from qna.llm import get_embeddings
+from qna.constants import CACHE_TYPE, INDEX_NAME, REDIS_URL
+
+def get_cache():
+    # construct cache implementation based on env var
+    if CACHE_TYPE == "semantic":
+        from langchain.cache import RedisSemanticCache
+        print("Using semantic cache")
+        embeddings = get_embeddings()
+        return RedisSemanticCache(
+            redis_url=REDIS_URL,
+            embedding=embeddings,
+            score_threshold=0.2
+        )
+    elif CACHE_TYPE == "standard":
+        from langchain.cache import RedisCache
+        return RedisCache(Redis.from_url(REDIS_URL))
+    return None
+
+
+
+def get_vectorstore(documents: List[Document]=None) -> "RedisVBD":
+    """Create the Redis vectorstore."""
+
+    embeddings = get_embeddings()
+
+    try:
+        vectorstore = RedisVDB.from_existing_index(
+            embedding=embeddings,
+            index_name=INDEX_NAME,
+            redis_url=REDIS_URL
+        )
+        return vectorstore
+    except:
+        pass
+
+    vectorstore = RedisVDB.from_documents(
+        documents=documents,
+        embedding=embeddings,
+        index_name=INDEX_NAME,
+        redis_url=REDIS_URL
+    )
+    return vectorstore
