+++ app/qna/prompt.py
@@ -1,0 +1,25 @@
+from langchain.prompts import PromptTemplate
+
+
+def basic_prompt():
+    # Define our prompt
+    prompt_template = """Use the following pieces of context to answer the question at the end. If you don't know the answer, say that you don't know, don't try to make up an answer.
+
+    This should be in the following format:
+
+    Question: [question here]
+    Answer: [answer here]
+
+    Begin!
+
+    Context:
+    ---------
+    {context}
+    ---------
+    Question: {question}
+    Answer:"""
+
+    prompt = PromptTemplate(
+        template=prompt_template,
+        input_variables=["context", "question"]
+    )
